Change Default Min Num Replicas for RF > 3 #5426
Summary:
Explicitly set the min_num_replicas for each AZ such that their sum equals the universe-wide RF. The number of replicas per AZ will be allocated such that the AZ with the most nodes will always have at least as many replicas as all other AZs, and at most one extra replica.

Simplify the allocation of AZs.

Also fixed a bug causing the placement warning text during universe creation to not show up. Added some extra logic to these warning messages to more intelligently be able to tell whether the proposed universe configuration is resilient to node failure, AZ failure, or whole region failure based on the allocation of replicas per AZ.

Removed hardcoded MAX_ZONES of 5, and instead dynamically limit the number of zones to the RF.

Changed shuffling of all AZs across all regions selected to be round-robin'd such that we ensure an even distribution of AZs across all regions + ensure that our placement code returns a deterministic set of AZs to use.

Test Plan:
Create different universes across different RF, # of AZ, and # of nodes to ensure that min_num_replicas are set properly. Verified that the appropriate warning text would show up depending on the universe configuration proposed.

For example:
  - RF=x, # AZ = 1, # nodes = y would result in min_num_replicas of x.
  - RF=5, # AZ = 3, # nodes = 6 with distribution of 3/1/2 would result in min_num_replicas of 2/1/2.
  - RF=3, # AZ = 3, # nodes = 3 with distribution of 1/1/1 would result in min_num_replicas of 1/1/1.
  - RF=5, # AZ = 4, # nodes = 5 with distribution of 2/1/1/1 would result in min_num_replicas of 2/1/1/1.
  - RF=7, # AZ = 6, # nodes = 16 with distribution of 5/4/3/2/1/1 would result in min_num_replicas of 2/1/1/1/1/1.

Also verified there was no regression with k8s placement through various deployment + cluster resizing tests.

Reviewers: bogdan, rahuldesirazu, arnav, sanketh

Reviewed By: sanketh

Differential Revision: https://phabricator.dev.yugabyte.com/D9570

+++ managed/src/main/java/com/yugabyte/yw/commissioner/tasks/RemoveNodeFromUniverse.java
@@ -126,7 +126,7 @@ public class RemoveNodeFromUniverse extends UniverseTaskBase {
         int rfInZone = PlacementInfoUtil.getZoneRF(pi, currentNode.cloudInfo.cloud,
                                                    currentNode.cloudInfo.region,
                                                    currentNode.cloudInfo.az);
-        int nodesInZone = PlacementInfoUtil.getNumActiveTserversInZone(
+        long nodesInZone = PlacementInfoUtil.getNumActiveTserversInZone(
             universe.getNodes(), currentNode.cloudInfo.cloud,
             currentNode.cloudInfo.region, currentNode.cloudInfo.az);
